lifeops-ai/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
apscheduler
httpx
pydantic
